Break out Simulator class into its own .cpp file
