+++ Makefile
@@ -1,5 +1,5 @@
 %.o: %.cpp
 	g++ -g -c -o $@ $<
 
-RoboSim.exe: MainSimRobo.o
-	g++ -g -o $@ $<
+RoboSim.exe: MainSimRobo.o Simulator.o
+	g++ -g -o $@ $^
